Drop pylons support
