+++ ckanext/dcat/plugins/__init__.py
@@ -6,12 +6,11 @@ import os
 from ckantoolkit import config
 
 from ckan import plugins as p
-try:
-    from ckan.lib.plugins import DefaultTranslation
-except ImportError:
-    class DefaultTranslation(object):
-        pass
 
+from ckan.lib.plugins import DefaultTranslation
+
+import ckanext.dcat.blueprints as blueprints
+import ckanext.dcat.cli as cli
 
 from ckanext.dcat.logic import (dcat_dataset_show,
                                 dcat_catalog_show,
@@ -21,8 +20,6 @@ from ckanext.dcat.logic import (dcat_dataset_show,
                                 )
 from ckanext.dcat import utils
 
-from ckanext.dcat.plugins.flask_plugin import MixinDCATPlugin, MixinDCATJSONInterface
-
 
 CUSTOM_ENDPOINT_CONFIG = 'ckanext.dcat.catalog_endpoint'
 TRANSLATE_KEYS_CONFIG = 'ckanext.dcat.translate_keys'
@@ -31,7 +28,7 @@ HERE = os.path.abspath(os.path.dirname(__file__))
 I18N_DIR = os.path.join(HERE, u"../i18n")
 
 
-class DCATPlugin(MixinDCATPlugin, p.SingletonPlugin, DefaultTranslation):
+class DCATPlugin(p.SingletonPlugin, DefaultTranslation):
 
     p.implements(p.IConfigurer, inherit=True)
     p.implements(p.ITemplateHelpers, inherit=True)
@@ -39,6 +36,18 @@ class DCATPlugin(MixinDCATPlugin, p.SingletonPlugin, DefaultTranslation):
     p.implements(p.IAuthFunctions, inherit=True)
     p.implements(p.IPackageController, inherit=True)
     p.implements(p.ITranslation, inherit=True)
+    p.implements(p.IClick)
+    p.implements(p.IBlueprint)
+
+    # IClick
+
+    def get_commands(self):
+        return cli.get_commands()
+
+    # IBlueprint
+
+    def get_blueprint(self):
+        return [blueprints.dcat]
 
     # ITranslation
 
@@ -123,9 +132,15 @@ class DCATPlugin(MixinDCATPlugin, p.SingletonPlugin, DefaultTranslation):
         return data_dict
 
 
-class DCATJSONInterface(MixinDCATJSONInterface, p.SingletonPlugin):
+class DCATJSONInterface(p.SingletonPlugin):
     p.implements(p.IActions)
     p.implements(p.IAuthFunctions, inherit=True)
+    p.implements(p.IBlueprint)
+
+    # IBlueprint
+
+    def get_blueprint(self):
+        return [blueprints.dcat_json_interface]
 
     # IActions
