+++ ckanext/dcat/processors.py
@@ -426,14 +426,6 @@ Operation mode.
 
     config.update({DCAT_EXPOSE_SUBCATALOGS: args.subcatalogs})
 
-    # Workaround until the core translation function defaults to the Flask one
-    from paste.registry import Registry
-    from ckan.lib.cli import MockTranslator
-    registry = Registry()
-    registry.prepare()
-    from pylons import translator
-    registry.register(translator, MockTranslator())
-
     if args.mode == 'produce':
         serializer = RDFSerializer(profiles=args.profile,
                                    compatibility_mode=args.compat_mode)
